+++ tests/builtins/strchr.c
@@ -255,7 +255,7 @@ void strchr_escaping() {
   CHAR_ARRAY(s,4);
   {
     int x;
-    *((int *)s) = &x;
+    *((int *)s) = (int) &x;
   }
   IF_NONDET(s[0], 0);
   STRCHR(RES, z1, s, 0, c); // alarm
